bot/requirements.txt
Pillow==7.2.0
PyQt5==5.15.0
PyQt5-sip==12.8.0
selenium==3.141.0
urllib3==1.25.9
